backend/requirements.txt
absl-py==1.2.0
asgiref==3.5.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
flatbuffers==2.0.7
gast==0.4.0
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
python-dotenv==0.21.0
idna==3.4
joblib==1.2.0
libclang==14.0.6
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.1
packaging==21.3
protobuf==3.19.5
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
termcolor==2.0.1
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.12
Werkzeug==2.2.2
win32-setctime==1.1.0
wrapt==1.14.1
